requirements.txt
matplotlib==3.10.0
numpy==2.2.3
openai==1.63.2
ordered-set==4.1.0
pandas==2.2.3
pillow==11.1.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
ragas==0.2.13
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
